Move functions and types as per issue #2961
